release: promote develop to main for v1.0.0

+++ backend/src/main/java/vn/nguongocso/certification/dto/response/ProductionLotTestCriteriaResponse.java
@@ -27,7 +27,7 @@ public class ProductionLotTestCriteriaResponse {
     @Builder
     public static class TestCriterionItemResponse {
         @JsonProperty("criteriaId")
-        private Integer criteriaId;
+        private Long criteriaId;
 
         @JsonProperty("code")
         private String code;
